rename cover.py to weighted.py
